added AppNavigation component
